refactor(http_server): re-internalize spawn+rollback+publish in add_activity_to_pane
Move PTY spawn, rollback-on-failure, and `publish_window_layout` back
into `AppState::add_activity_to_pane` so that all layout broadcasts
originate exclusively from AppState methods. The handler is now thin:
parse → delegate → return. The now-dead standalone `publish_window_layout`
free function in `handlers.rs` is removed.
